+++ src/main/java/no/nav/familie/dokumentgenerator/demo/utils/GenerateUtils.java
@@ -1,0 +1,76 @@
+package no.nav.familie.dokumentgenerator.demo.utils;
+
+import org.commonmark.node.Node;
+import org.commonmark.parser.Parser;
+import org.commonmark.renderer.html.HtmlRenderer;
+import org.jsoup.Jsoup;
+import org.jsoup.nodes.Document;
+import org.jsoup.nodes.Element;
+
+import java.io.IOException;
+import java.net.URI;
+import java.net.http.HttpClient;
+import java.net.http.HttpRequest;
+import java.net.http.HttpResponse;
+
+public class GenerateUtils {
+    private String pdfGenURl = "http://localhost:8090/api/v1/genpdf/html/";
+
+    private String getPdfGenURl() {
+        return pdfGenURl;
+    }
+
+    private Document appendHtmlMetadata(String html) {
+
+        Document document = Jsoup.parse(html);
+        Element head = document.head();
+        String css = FileUtils.getCssFile("main.css");
+
+        head.append("<meta charset=\"UTF-8\">");
+        head.append("\n<style>\n" + css + "\n</style>");
+
+        return document;
+    }
+
+    public String convertMarkdownTemplateToHtml(String content) {
+        Node document = parseDocument(content);
+        String html = renderToHTML(document);
+        Document htmlDocument = appendHtmlMetadata(html);
+        return htmlDocument.html();
+    }
+
+    public byte[] generatePDF(String html, String applicationName) {
+        HttpClient client = HttpClient.newHttpClient();
+        HttpRequest request = HttpRequest.newBuilder()
+                .uri(URI.create(getPdfGenURl() + applicationName))
+                .header("Content-Type", "text/html;charset=UTF-8")
+                .POST(HttpRequest.BodyPublishers.ofString(html))
+                .build();
+
+        try {
+            HttpResponse<byte[]> response = client.send(request, HttpResponse.BodyHandlers.ofByteArray());
+            if (response.statusCode() == 200) {
+                return response.body();
+            }
+        } catch (IOException | InterruptedException e) {
+            e.printStackTrace();
+        }
+        return  null;
+    }
+
+    private Node parseDocument(String content) {
+        return getMarkdownToHtmlParser().parse(content);
+    }
+
+    private String renderToHTML(Node document) {
+        return getHtmlRenderer().render(document);
+    }
+
+    private Parser getMarkdownToHtmlParser() {
+        return Parser.builder().build();
+    }
+
+    private HtmlRenderer getHtmlRenderer() {
+        return HtmlRenderer.builder().build();
+    }
+}
