+++ src/main/java/no/nav/familie/dokumentgenerator/demo/utils/JsonUtils.java
@@ -1,0 +1,35 @@
+package no.nav.familie.dokumentgenerator.demo.utils;
+
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+import java.io.File;
+import java.io.IOException;
+import java.net.URISyntaxException;
+import java.net.URL;
+
+public class JsonUtils {
+
+    public JsonNode getJsonFromString(String json) throws IOException {
+        ObjectMapper mapper = new ObjectMapper();
+        return mapper.readTree(json);
+    }
+
+    private JsonNode readJsonFile(URL path) {
+        if (path != null) {
+            ObjectMapper mapper = new ObjectMapper();
+            try {
+                return mapper.readTree(new File(path.toURI()));
+            } catch (IOException | URISyntaxException e) {
+                System.out.println("Kan ikke finne JSON fil!");
+                e.printStackTrace();
+            }
+        }
+        return null;
+    }
+
+    public JsonNode getTestSetField(String templateName, String testSet){
+        URL path = ClassLoader.getSystemResource("templates/" + templateName + "/testdata/" + testSet + ".json");
+        return readJsonFile(path);
+    }
+}
